backend/requirements.txt
fastapi
uvicorn
web3
requests
pydantic
python-dotenv

